wireguard: tweaked an ifdef to get things compiling
Kinda odd that this is happening on .77 of an LTS kernel...  but
wireguard stopped comping due to a redefinition of a preprocessor
macro.  Said definition was already in an if block checking for v5.5,
so I changed it to v5.4.  Things compile now... but I'm not 100% sure
that's the right fix.  I should make sure to test wg a little.